// archive/packages/dataforge-old/src/generated/client-entities.ts
export interface User {
  id: any;
  createdAt: any;
  updatedAt: any;
  name: any;
  email?: any;
  emailVerified: any;
  image?: any;
  isSuperAdmin: any;
  role: any;
  accounts: any;
  sessions: any;
  assignedTasks: any;
  comments: any;
  ownedProjects: any;
}

export interface task_tags {
  Task_owner: any;
  Tag_inverse: any;
}

export interface Task {
  id: any;
  createdAt: any;
  updatedAt: any;
  clientId?: any;
  title: any;
  description?: any;
  legacyStatus?: any;
  priority: any;
  dueDate?: any;
  startDate?: any;
  completedAt?: any;
  timeRange?: any;
  estimatedDuration?: any;
  legacyTags?: any;
  project?: any;
  assignee?: any;
  comments: any;
  tags: any;
}
// ...
export class User implements User {}
export class task_tags implements task_tags {}
export class Task implements Task {}